requirements.txt
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
greenlet==3.5.0
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
matplotlib==3.10.9
matplotlib-inline==0.2.1
narwhals==2.20.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==26.2
pandas==2.2.2
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.12
pure_eval==0.2.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.49
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.2
wcwidth==0.6.0
